feat: add command .qsolo unrated/rated
