+++ src/solo3v3_loader.cpp
@@ -16,8 +16,11 @@
  */
 
 void AddSC_Solo_3v3_Arena();
+void AddSC_Solo_3v3_commandscript();
+
 
 void Addmod_arena_3v3_solo_queueScripts()
 {
     AddSC_Solo_3v3_Arena();
+    AddSC_Solo_3v3_commandscript();
 }
